+++ app/src/main/res/layout/fragment_modify_messagenickname.xml
@@ -1,0 +1,151 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    tools:context=".mypage.ModifyMessageNickname">
+
+    <TextView
+        android:id="@+id/know_mypage"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="Modify"
+        android:textSize="25dp"
+        android:textStyle="bold"
+        android:layout_marginLeft="15dp"
+        android:layout_marginTop="10dp"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:ignore="MissingConstraints" />
+
+    <View
+        android:id="@+id/line"
+        android:layout_width="match_parent"
+        android:layout_height="2dp"
+        android:layout_margin="10dp"
+        android:background="#CCCCCC"
+        app:layout_constraintTop_toBottomOf="@+id/know_mypage" />
+
+    <LinearLayout
+        android:id="@+id/mess_linear"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:layout_marginLeft="20dp"
+        android:layout_marginRight="20dp"
+        android:layout_marginTop="30dp"
+        android:background="@drawable/home_yellow_shape"
+        app:layout_constraintTop_toBottomOf="@id/line">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="현재 응원메시지"
+            android:textStyle="bold"
+            android:layout_margin="10dp"
+            android:textSize="20dp" />
+
+        <TextView
+            android:id="@+id/text_mess"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="10dp"
+            android:text="월급은 스쳐가는 바람일 뿐."
+            android:layout_margin="10dp"
+            android:textAlignment="viewEnd"
+            android:textSize="20dp" />
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="변경 할 응원메시지"
+            android:layout_margin="10dp"
+            android:layout_marginTop="30dp"
+            android:textStyle="bold"
+            android:textSize="20dp" />
+
+        <EditText
+            android:id="@+id/modify_message"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="15dp"
+            android:hint="20자 이내로 작성해주세요."
+            android:layout_margin="10dp"
+            android:maxLength="20"/>
+
+    </LinearLayout>
+
+    <Button
+        android:id="@+id/registerMessBtn"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="변경"
+        android:textSize="20dp"
+        android:backgroundTint="#FAED7D"
+        android:textColor="@color/black"
+        android:layout_marginTop="10dp"
+        app:layout_constraintTop_toBottomOf="@+id/mess_linear"
+        app:layout_constraintRight_toRightOf="@+id/mess_linear"/>
+
+    <LinearLayout
+        android:id="@+id/nick_linear"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:layout_marginLeft="20dp"
+        android:layout_marginRight="20dp"
+        android:layout_marginTop="30dp"
+        android:background="@drawable/home_green_shape"
+        app:layout_constraintTop_toBottomOf="@id/registerMessBtn">
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="현재 닉네임"
+            android:textStyle="bold"
+            android:layout_margin="10dp"
+            android:textSize="20dp" />
+
+        <TextView
+            android:id="@+id/text_nick"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="10dp"
+            android:text="Clova"
+            android:layout_margin="10dp"
+            android:textAlignment="viewEnd"
+            android:textSize="20dp" />
+
+        <TextView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="변경 할 닉네임"
+            android:layout_margin="10dp"
+            android:layout_marginTop="30dp"
+            android:textStyle="bold"
+            android:textSize="20dp" />
+
+        <EditText
+            android:id="@+id/modify_nickname"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:hint="5자 이내로 작성해주세요."
+            android:layout_margin="10dp"
+            android:maxLength="5"/>
+
+    </LinearLayout>
+
+    <Button
+        android:id="@+id/registerNickBtn"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="변경"
+        android:textSize="20dp"
+        android:backgroundTint="#CEF279"
+        android:textColor="@color/black"
+        android:layout_marginTop="10dp"
+        app:layout_constraintTop_toBottomOf="@+id/nick_linear"
+        app:layout_constraintRight_toRightOf="@+id/nick_linear"/>
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
